backend/pyproject.toml
[project]
name = "kb-manager"
version = "0.1.0"
description = "KB Manager v2 Backend — AI-powered content ingestion service"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "pydantic-settings>=2.0.0",
    "boto3>=1.35.0",
    "aiobotocore>=2.15.0",
    "httpx>=0.27.0",
    "strands-agents>=0.1.0",
]

[project.optional-dependencies]
dev = [
    "hypothesis>=6.100.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx",
]

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["kb_manager*"]
